refactor: standardize all game page.tsx files (#112)
- Extract game logic out of page.tsx for 3 non-standard games:
  - memory/page.tsx -> MemoryClient.tsx (was 'use client' with hooks)
  - puzzles/page.tsx -> PuzzlesClient.tsx (was 'use client' with useState)
  - building/page.tsx -> BuildingGameClient.tsx (was 'use client' with useEffect)

- All 3 pages now follow the standard 4-line pattern:
    import *Client + generateGameMetadata + metadata export + Page()

- Added missing metadata (SEO) to memory, puzzles, and building pages

- Fix reflex/page.tsx: remove extra blank lines (cosmetic)

All 40 game page.tsx files now follow identical structure.
71/71 tests pass.
