+++ gamesformykids/app/games/reflex/page.tsx
@@ -1,12 +1,6 @@
 ﻿import ReflexGameClient from './ReflexGameClient';
 import { generateGameMetadata } from '@/lib/utils/game/gameMetadata';
 
-
-
-
-
 export const metadata = generateGameMetadata('reflex');
 
-export default function ReflexPage() {
-  return <ReflexGameClient />;
-}
+export default function ReflexPage() { return <ReflexGameClient />; }
